MET-80: Use ts instead of tsx
