+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/shared/hooks/use-focus.ts
@@ -1,0 +1,14 @@
+import {useRef, useCallback, useEffect, RefObject} from 'react';
+
+const useFocus = (): [RefObject<HTMLInputElement>, () => void] => {
+  const ref = useRef<HTMLInputElement | null>(null);
+  const setFocus = useCallback(() => {
+    if (ref.current !== null) ref.current.focus();
+  }, [ref]);
+
+  useEffect(setFocus, []);
+
+  return [ref, setFocus];
+};
+
+export {useFocus};
