+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/shared/hooks/use-shortcut.ts
@@ -1,4 +1,4 @@
-import React, {useCallback} from 'react';
+import {useCallback, useEffect} from 'react';
 import {Key} from 'akeneomeasure/shared/key';
 
 const useShortcut = (key: Key, callback: () => void) => {
@@ -7,7 +7,7 @@ const useShortcut = (key: Key, callback: () => void) => {
     callback,
   ]);
 
-  React.useEffect(() => {
+  useEffect(() => {
     document.addEventListener('keydown', memoizedCallback);
     return () => document.removeEventListener('keydown', memoizedCallback);
   }, [memoizedCallback]);
